Add a python implementation of structured light, some test videos
