+++ src/pathfinder/.gitignore
@@ -1,0 +1,2 @@
+motion
+calib
